Add seeds to match vanilla behavior
These plants in vanilla are able to use their fruits as their seeds,
in contrast to wheat, melons, pumpkins, and beets. Having the AgriCraft
defaults behave differently was simply confusing. The mutations for thm
are still useful, like when a player can't find a particular plant but
does have others that they can cross breed.